+++ CHANGELOG.md
@@ -14,12 +14,14 @@
 ### Others
 
 * [#8] - Rename Renovate configuration file to dot file
+* [#9] - Include .js files in ESLint target
 
 [Unreleased]: https://github.com/sounisi5011/readme-generator/compare/v0.0.1...master
 [#5]: https://github.com/sounisi5011/readme-generator/pull/5
 [#6]: https://github.com/sounisi5011/readme-generator/pull/6
 [#7]: https://github.com/sounisi5011/readme-generator/pull/7
 [#8]: https://github.com/sounisi5011/readme-generator/pull/8
+[#9]: https://github.com/sounisi5011/readme-generator/pull/9
 
 ## [0.0.1] - 2020-05-24 UTC